Implement adding to cart logic
This involved creating some helper methods to update the cart.

First, we must find if the cart already contains the item id, if it does, then we update the quantity.
If it doesn't, then a new array element is created.

An important point above is that the cart must be updated with a NEW reference so that the cartItems derived from the cart is updated. If the same reference is passed, it may not detect a change and re-render the App component.