Write test suite for Shop component
This was a simple test suite with two test cases.

The first case tests the rendered output via snapshot, and the second test tests to ensure that no Item Cards were created, and thus, the shop container has no children.

The second test used the getByTestId query from the React Testing Library, and I had to add a data-testid attribute to the shop container div.