Write test suite for App.jsx
The biggest challenges for this test suite was figuring out how to approach the BrowserRouter and the fetch API. After some research, it appeared that MemoryRouter was the correct approach, which was implemented easily.

To intercept the fetch API requests, I used the msw package which I used to configure the handler to return a simplified version of the products array for testing purposes.

Finally, I tested the bigger picture functionalities, adding and removing items to and from the cart.